[metasrv] refactor: impl KVApi for MetaNode, as it already provides the methods in KVApi
